+++ command/scp_command.go
@@ -2,6 +2,8 @@ package command
 
 import (
 	"fmt"
+	"github.com/10gen-labs/slogger/v1"
+	"github.com/evergreen-ci/evergreen"
 	"io"
 	"os/exec"
 )
@@ -65,5 +67,9 @@ func (self *ScpCommand) Start() error {
 }
 
 func (self *ScpCommand) Stop() error {
-	return self.Cmd.Process.Kill()
+	if self.Cmd != nil && self.Cmd.Process != nil {
+		return self.Cmd.Process.Kill()
+	}
+	evergreen.Logger.Logf(slogger.WARN, "Trying to stop command but Cmd / Process was nil")
+	return nil
 }
